Add function file src/lib/date.ts
