+++ src/components/todo/WeeksTodoLists.tsx
@@ -8,6 +8,7 @@ import {
 import { AddTodo, TodoList } from './index';
 import { State } from '../../reducks/store/types';
 import { TodoLists } from '../../reducks/todoLists/types';
+import { changePrevDateType, getStartDate, getWeekDay } from '../../lib/date';
 
 interface ExistTodoListsProps {
   selectedDate: Date;
@@ -19,22 +20,8 @@ const WeeksTodoLists = (props: ExistTodoListsProps) => {
   const dueTodoLists = getDueTodoLists(selector);
   const todoListsMessage = getTodoListsMessage(selector);
   const dt: Date = props.selectedDate !== null ? props.selectedDate : new Date();
-  const weekdays = ['日', '月', '火', '水', '木', '金', '土'];
   const selectedDate = new Date(dt);
-  const selectedWeekDay = selectedDate.getDay();
-  const _startDate = new Date(
-    selectedDate.getFullYear(),
-    selectedDate.getMonth(),
-    selectedDate.getDate() - selectedWeekDay
-  );
-
-  const changePrevDateType = (date: string) => {
-    const prevDates = date.split(/[/()]/, 3);
-    const prevYear = Number(prevDates[0]);
-    const prevMonth = Number(prevDates[1]) - 1;
-    const prevDate = Number(prevDates[2]);
-    return new Date(prevYear, prevMonth, prevDate);
-  };
+  const _startDate = getStartDate(selectedDate);
 
   const week = useCallback(
     (todoLists: TodoLists) => {
@@ -54,7 +41,6 @@ const WeeksTodoLists = (props: ExistTodoListsProps) => {
               ? changePrevDateType(todoList.implementation_date).getTime()
               : changePrevDateType(todoList.due_date).getTime();
           const weekDate: number = date.getTime();
-
           if (prevDate === weekDate) {
             dateTodoLists.push(<TodoList todoListItem={todoList} key={todoList.id} />);
           }
@@ -63,7 +49,7 @@ const WeeksTodoLists = (props: ExistTodoListsProps) => {
         week.push(
           <div key={date.getDate()}>
             <p>
-              {date.getMonth() + 1}/{date.getDate()} （{weekdays[date.getDay()]}）
+              {date.getMonth() + 1}/{date.getDate()} （{getWeekDay(date)}）
             </p>
             {dateTodoLists}
             <AddTodo />
